+++ src/main/java/com/gdg/z_meet/domain/booth/entity/Category.java
@@ -1,0 +1,5 @@
+package com.gdg.z_meet.domain.booth.entity;
+
+public enum Category {
+    DATE, FOOD, GOODS, EVENT
+}
