+++ src/main/java/com/gdg/z_meet/domain/chat/entity/ChatRoom.java
@@ -1,0 +1,21 @@
+package com.gdg.z_meet.domain.chat.entity;
+
+import com.gdg.z_meet.global.common.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class ChatRoom extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "chat_room_id", unique = true)
+    private Long id;
+
+    @Column(nullable = false)
+    private String name;
+}
