+++ src/main/java/com/gdg/z_meet/domain/chat/entity/JoinChat.java
@@ -1,0 +1,27 @@
+package com.gdg.z_meet.domain.chat.entity;
+
+import com.gdg.z_meet.domain.user.entity.User;
+import com.gdg.z_meet.global.common.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class JoinChat extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "join_chat_id", unique = true)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "chat_room_id")
+    private ChatRoom chatRoom;
+}
