+++ src/main/java/com/gdg/z_meet/domain/chat/entity/Message.java
@@ -1,0 +1,32 @@
+package com.gdg.z_meet.domain.chat.entity;
+
+import com.gdg.z_meet.domain.user.entity.User;
+import com.gdg.z_meet.global.common.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class Message extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "message_id", unique = true)
+    private Long id;
+
+    @Column(nullable = false)
+    private String content;
+
+    private Boolean isRead;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "chat_room_id")
+    private ChatRoom chatRoom;
+}
