+++ src/main/java/com/gdg/z_meet/domain/meeting/entity/TeamType.java
@@ -1,0 +1,5 @@
+package com.gdg.z_meet.domain.meeting.entity;
+
+public enum TeamType {
+    TWO_TO_TWO, THREE_TO_THREE
+}
