+++ src/main/java/com/gdg/z_meet/domain/order/entity/Order.java
@@ -1,0 +1,27 @@
+package com.gdg.z_meet.domain.order.entity;
+
+import com.gdg.z_meet.domain.user.entity.User;
+import com.gdg.z_meet.global.common.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Table(name="orders")
+public class Order extends BaseEntity {
+
+    @Id
+    @Column(name = "order_id", unique = true)
+    private String id;            // UUID
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "toss_payments_id")
+    private TossPayments tossPayments;
+}
