+++ src/main/java/com/gdg/z_meet/domain/order/entity/TossPaymentStatus.java
@@ -1,0 +1,6 @@
+package com.gdg.z_meet.domain.order.entity;
+
+public enum TossPaymentStatus {
+    ABORTED, CANCELED, DONE , EXPIRED, IN_PROGRESS,
+    PARTIAL_CANCELED, READY, WAITING_FOR_DEPOSIT
+}
