+++ src/main/java/com/gdg/z_meet/domain/user/dto/Token.java
@@ -1,0 +1,16 @@
+package com.gdg.z_meet.domain.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Token {
+    private String accessToken;
+    private String refreshToken;
+    private String key;
+}
