+++ src/main/java/com/gdg/z_meet/domain/user/entity/User.java
@@ -1,0 +1,71 @@
+package com.gdg.z_meet.domain.user.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name="user")
+public class User implements UserDetails {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="user_id", unique = true, nullable = false)
+    private Long id;
+
+    @Column(unique = true, nullable = false)
+    private String studentNumber;
+
+    @Column(unique = true, nullable = false)
+    private String password;
+
+    @Column(unique = true, nullable = false)
+    private String name;
+
+//    @OneToOne(mappedBy = "user", cascade = CascadeType.ALL)
+//    private UserProfile userProfile;
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return null;
+    }
+
+    @Override
+    public String getPassword() {
+        return password;
+    }
+
+    @Override
+    public String getUsername() {
+        return studentNumber;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+}
