+++ src/main/java/com/gdg/z_meet/domain/user/entity/enums/Grade.java
@@ -1,0 +1,5 @@
+package com.gdg.z_meet.domain.user.entity.enums;
+
+public enum Grade {
+    FIRST, SECOND, THIRD, FOURTH
+}
