+++ src/main/java/com/gdg/z_meet/domain/user/entity/enums/IdealAge.java
@@ -1,0 +1,5 @@
+package com.gdg.z_meet.domain.user.entity.enums;
+
+public enum IdealAge {
+    OLDER, YOUNGER, SAME, NO_MATTER
+}
