+++ src/main/java/com/gdg/z_meet/domain/user/entity/enums/IdealType.java
@@ -1,0 +1,6 @@
+package com.gdg.z_meet.domain.user.entity.enums;
+
+public enum IdealType {
+    DOG, CAT, HAMSTER, FOX,
+    BEAR, DINOSAUR, RABBIT, WOLF
+}
