+++ src/main/java/com/gdg/z_meet/domain/user/entity/enums/MBTI.java
@@ -1,0 +1,6 @@
+package com.gdg.z_meet.domain.user.entity.enums;
+
+public enum MBTI {
+    INTJ, INTP, ENTJ, ENTP, INFJ, INFP, ISTP, ISFP,
+    ENFJ, ENFP, ISTJ, ISFJ, ESTJ, ESFJ, ESTP, ESFP;
+}
