+++ src/main/java/com/gdg/z_meet/domain/user/entity/enums/Music.java
@@ -1,0 +1,6 @@
+package com.gdg.z_meet.domain.user.entity.enums;
+
+public enum Music {
+    KPOP, POP, BALLOD, HIPHOP,
+    JPOP, ZAZZ, BAND, CLASSICAL
+}
