+++ src/main/java/com/gdg/z_meet/domain/user/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.gdg.z_meet.domain.user.repository;
+
+import com.gdg.z_meet.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByStudentNumber(String studentNumber);
+}
