+++ src/main/java/com/gdg/z_meet/domain/user/service/UserDetailsService.java
@@ -1,0 +1,22 @@
+package com.gdg.z_meet.domain.user.service;
+
+import com.gdg.z_meet.domain.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class UserDetailsService implements org.springframework.security.core.userdetails.UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    @Transactional
+    public UserDetails loadUserByUsername(String studentNumber) throws UsernameNotFoundException {
+        return userRepository.findByStudentNumber(studentNumber)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found with student number: " + studentNumber));
+    }
+}
