--- src/main/java/com/gdg/z_meet/user/dto/Token.java
@@ -1,16 +1,0 @@
-package com.gdg.z_meet.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Token {
-    private String accessToken;
-    private String refreshToken;
-    private String key;
-}
