--- src/main/java/com/gdg/z_meet/user/dto/UserRes.java
@@ -1,14 +1,0 @@
-package com.gdg.z_meet.user.dto;
-
-import lombok.*;
-
-public class UserRes {
-    @Builder
-    @NoArgsConstructor
-    @AllArgsConstructor
-    @Getter
-    @Setter
-    public static class SignUpRes {
-        private String message;
-    }
-}
