--- src/main/java/com/gdg/z_meet/user/entity/profile/Emoji.java
@@ -1,5 +1,0 @@
-package com.gdg.z_meet.user.entity.profile;
-
-public enum Emoji{
-    DIAMOND
-}
