--- src/main/java/com/gdg/z_meet/user/entity/profile/Gender.java
@@ -1,5 +1,0 @@
-package com.gdg.z_meet.user.entity.profile;
-
-public enum Gender {
-    MAN, WOMAN
-}
