--- src/main/java/com/gdg/z_meet/user/entity/profile/IdealType.java
@@ -1,6 +1,0 @@
-package com.gdg.z_meet.user.entity.profile;
-
-public enum IdealType{
-    DOG, CAT, HAMSTER, FOX,
-    BEAR, DINOSAUR, RABBIT, WOLF
-}
