--- src/main/java/com/gdg/z_meet/user/entity/profile/MBTI.java
@@ -1,6 +1,0 @@
-package com.gdg.z_meet.user.entity.profile;
-
-public enum MBTI {
-    INTJ, INTP, ENTJ, ENTP, INFJ, INFP, ISTP, ISFP,
-    ENFJ, ENFP, ISTJ, ISFJ, ESTJ, ESFJ, ESTP, ESFP;
-}
