--- src/main/java/com/gdg/z_meet/user/entity/profile/Music.java
@@ -1,6 +1,0 @@
-package com.gdg.z_meet.user.entity.profile;
-
-public enum Music{
-    KPOP, POP, BALLOD, HIPHOP,
-    JPOP, ZAZZ, BAND, CLASSICAL
-}
