--- src/main/java/com/gdg/z_meet/user/entity/profile/Style.java
@@ -1,6 +1,0 @@
-package com.gdg.z_meet.user.entity.profile;
-
-public enum Style{
-    CASUAL, STREET, NEAT, VINTAGE,
-    HEAP, SPORTY, FORMAL, CUTIE
-}
